config/lru_test.go
/*
Copyright 2013 Google Inc.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

     http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

package config

import (
	"fmt"
	"testing"

	"github.com/corestoreio/pkg/sync/bgwork"
)

var lruGetTests = []struct {
	name       string
	keyToAdd   *Path
	keyToGet   *Path
	expectedOk bool
}{
	{"01 hit", MustMakePath("aa/bb/cc"), MustMakePath("aa/bb/cc"), true},
	{"02 miss", MustMakePath("aa/bb/cc"), MustMakePath("aa/bb/cc").BindStore(11), false},
	{"03 hit", MustMakePath("aa/bb/cc").BindWebsite(3), MustMakePath("aa/bb/cc").BindWebsite(3), true},
	{"04 miss", MustMakePath("aa/bb/cc").BindStore(44), MustMakePath("aa/bb/cc").BindWebsite(4), false},
	{"05 miss", MustMakePath("aa/bb/cc").BindStore(55), MustMakePath("aa/bb/cc").BindStore(56), false},
	{"06 hit", MustMakePath("aa/bb/cc").BindStore(6), MustMakePath("aa/bb/cc").BindStore(6), true},
}

var testLRUVal = MakeValue([]byte(`hello world`))

func TestLRUGet(t *testing.T) {

	newDistinctLRUSize := func(size int) func(*testing.T) {
		return func(t *testing.T) {
			for _, tt := range lruGetTests {
				lru := newLRU(size)
				lru.Add(tt.keyToAdd, testLRUVal)
				val, ok := lru.Get(tt.keyToGet)
				if ok != tt.expectedOk {
					t.Fatalf("%q %s: cache hit = %v; want %v", t.Name(), tt.name, ok, !ok)
				} else if ok && !val.equalData(testLRUVal) {
					t.Fatalf("Distinct: %q %s expected get to return %q but got %q", t.Name(), tt.name, testLRUVal.String(), val.String())
				}
			}
		}
	}

	newCommonLRUSize := func(size int) func(*testing.T) {
		return func(t *testing.T) {
			lru := newLRU(size)
			for _, tt := range lruGetTests {
				lru.Add(tt.keyToAdd, testLRUVal)
				val, ok := lru.Get(tt.keyToGet)
				if ok != tt.expectedOk {
					t.Fatalf("%q %s: cache hit = %v; want %v", t.Name(), tt.name, ok, !ok)
				} else if ok && !val.equalData(testLRUVal) {
					t.Fatalf("Common: %q %s expected get to return %q but got %q", t.Name(), tt.name, testLRUVal.String(), val.String())
				}
			}
		}
	}

	t.Run("distinct LRU size=0", newDistinctLRUSize(0))
	t.Run("distinct LRU size=1", newDistinctLRUSize(1))
	t.Run("common LRU size=0", newCommonLRUSize(0))
	t.Run("common LRU size=1", newCommonLRUSize(1))
}

func TestLRURemove(t *testing.T) {
	lru := newLRU(0)
	p := MustMakePath("gg/hh/ii").BindStore(33)
	lru.Add(p, testLRUVal)
	if val, ok := lru.Get(p); !ok {
		t.Fatal("TestRemove returned no match")
	} else if !val.equalData(testLRUVal) {
		t.Fatalf("TestRemove failed.  Expected %s, got %s", testLRUVal, val)
	}

	lru.Remove(p)
	if _, ok := lru.Get(p); ok {
		t.Fatal("TestRemove returned a removed entry")
	}
}

func TestLRUNew_Parallel(t *testing.T) {

	t.Run("add in parallel", func(t *testing.T) {
		lru := newLRU(2)
		bgwork.Wait(len(lruGetTests), func(idx int) {
			defer func() {
				if r := recover(); r != nil {
					t.Error(r)
				}
			}()
			tt := lruGetTests[idx]

			/* TODO THIS TEST IS FLAKY, with "-race" it passes. it fails depending on the LRU size. <= 3 fails, and 0 or > 3 succeeds always.
			$ go test -run=TestLRU -count=34
			panic: 06 hit: cache hit = false; want true: add:"stores/6/aa/bb/cc" get:"stores/6/aa/bb/cc"
			AFAIK problems with cache eviction
			*/
			lru.Add(tt.keyToAdd, testLRUVal)
			lru.Add(tt.keyToAdd, testLRUVal) // this is an ugly fix, but not that flaky anymore when running test with -count=40

			if val, ok := lru.Get(tt.keyToGet); ok != tt.expectedOk {
				panic(fmt.Sprintf("%s: cache hit = %v; want %v: add:%q get:%q", tt.name, ok, !ok, tt.keyToAdd, tt.keyToGet))
			} else if ok && !val.equalData(testLRUVal) {
				panic(fmt.Sprintf("%s expected get to return %s but got %v", tt.name, testLRUVal, val))
			}
		})
	})

	t.Run("add before parallel", func(t *testing.T) {
		lru := newLRU(5)

		for _, tt := range lruGetTests {
			lru.Add(tt.keyToAdd, testLRUVal)
		}

		bgwork.Wait(len(lruGetTests), func(idx int) {
			tt := lruGetTests[idx]

			val, ok := lru.Get(tt.keyToGet)
			if ok != tt.expectedOk {
				panic(fmt.Sprintf("%s: cache hit = %v; want %v", tt.name, ok, tt.expectedOk))
			} else if ok && !val.equalData(testLRUVal) {
				panic(fmt.Sprintf("%s expected get to return %s but got %v", tt.name, testLRUVal, val))
			}
		})
	})

}

func BenchmarkLRUNew_Parallel(b *testing.B) {
	const cacheSize = 5
	b.Run("single", func(b *testing.B) {
		lru := newLRU(cacheSize)
		for _, tt := range lruGetTests {
			lru.Add(tt.keyToAdd, testLRUVal)
		}
		b.ResetTimer()
		for i := 0; i < b.N; i++ {
			for _, tt := range lruGetTests {
				_, ok := lru.Get(tt.keyToGet)
				if ok != tt.expectedOk {
					b.Fatalf("%s: cache hit = %v; want %v", tt.name, ok, !ok)
				}
			}
		}
	})
	b.Run("parallel", func(b *testing.B) {
		lru := newLRU(cacheSize)
		for _, tt := range lruGetTests {
			lru.Add(tt.keyToAdd, testLRUVal)
		}

		b.ResetTimer()
		b.RunParallel(func(pb *testing.PB) {
			var val2 Value
			for pb.Next() {
				for _, tt := range lruGetTests {
					val, ok := lru.Get(tt.keyToGet)
					if ok != tt.expectedOk {
						b.Fatalf("%s: cache hit = %v; want %v", tt.name, ok, !ok)
					}
					val2 = val
				}
			}
			_ = val2 // ;-)
		})
	})

}
